cleaned up merge
