+++ src/App.css
@@ -68,3 +68,20 @@ ul{
     transform: rotate(360deg);
   }
 }
+.char-images {
+  height: 150px;
+  width: 150px;
+}
+
+.characters-container {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  gap: 20px;
+  background-image: ;
+
+}
+
+.imgContainer  {
+  object-fit: contain
+}
