+++ src/screens/Characters.jsx
@@ -1,31 +1,30 @@
-import React from 'react'
-import {useEffect, useState} from 'react'
-import {getCharacters} from '../../src/services/characters.js'
-import Character from '../components/Character.jsx'
+import { useState, useEffect } from 'react'
+import { getCharacters } from '../services/characters.js'
+import Character from '../components/Character.jsx';
 
 function Characters() {
-    const [characters , setCharacters] = useState([])
+  const [characters, setCharacters] = useState([])
 
-    useEffect(()=>{
-      fetchCharacters()
-    }, []) //only fire this function one time 
-  
-    async function fetchCharacters () {
-      const allCharacters = await getCharacters()
-      setCharacters(allCharacters)
-    }
-  
-    return (
-      <div>
-        <h1>All the characters!</h1>
-        <div className="characters-container">
-          {
-            characters.map((character)=>(
-              <Character character={character} />
-            ))
-          }
-        </div>
-      </div>
-    )
+  useEffect(() => {
+    fetchCharacters()
+  }, [])
+
+  async function fetchCharacters() {
+    const allCharacters = await getCharacters()
+    setCharacters(allCharacters)
   }
+
+  return (
+    <div className='pageContainer'>
+      <h1>All the Characters!</h1>
+      <div className="characters-container">
+        {characters?.map((character, idx) => (
+          <Character character={character} key={idx}/>
+        ))}
+      </div>
+    </div>
+  )
+}
+
 export default Characters
+
