--- src/screens/characters/characters.css
@@ -1,16 +1,0 @@
-.char-images {
-    height: 150px;
-    width: 150px;
-}
-
-.characters-container {
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: row;
-    gap: 20px;
-
-}
-
-.imgContainer  {
-object-fit: contain
-}
